% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{greta.dynamics}
\alias{greta.dynamics}
\title{greta.dynamics: a greta extension for modelling dynamical systems}
\description{
an extension to \href{https://greta-stats.org/}{greta}
with functions for simulating dynamical systems, defined by of ordinary
differential equations (see \code{\link[=ode_solve]{ode_solve()}}) or transition matrices
(\code{\link[=iterate_matrix]{iterate_matrix()}}).
}
