% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate.R
\name{iterate_matrix}
\alias{iterate_matrix}
\title{iterate transition matrices}
\usage{
iterate_matrix(matrix, initial_state = rep(1, ncol(matrix)),
  niter = 100)
}
\arguments{
\item{matrix}{either a square 2D transition matrix (with dimensions m x m),
or a 3D array (with dimensions n x m x m), giving one or more transition
matrices to iterate}

\item{initial_state}{either a column vector (with m elements) or a 3D array
(with dimensions n x m x 1) giving one or more initial states from which to
iterate the matrix}

\item{niter}{a positive integer giving the number of times to iterate the
matrix}
}
\value{
a named list with two greta arrays: \code{lambda} a scalar or vector
  giving the ratio of the first stage values between the final two
  iterations, and \code{final_state} a vector or matrix (with the same
  dimensions as \code{initial_state}) giving the state after the final
  iteration. If the system has convreged in \code{niter} iterations, these
  correspond to the asymptotic growth rate and stable stage distribution
  respectively.
}
\description{
calculate the intrinsic growth rate(s) and stable stage
  distribution(s) for a stage-structured dynamical system, encoded as a
  transition matrix
}
\details{
\code{iterate_matrix} can either act on single transition matrix and
  initial state (if \code{matrix} is 2D and \code{initial_state} is a column
  vector), or it can simultaneously act on \emph{n} different matrices and/or
  \emph{n} different initial states (if \code{matrix} and
  \code{initial_state} are 3D arrays). In the latter case, the first
  dimension of both objects should be the batch dimension \emph{n}.
}
\examples{


}
